Improve motor end-stop: stall (3s ~0A) or 0%/100%, then IDLE 5s arrive; MQTT gate until connected.

+++ 03.software/02.source_code/04.controller_unit_motor_freertos/system/drivers/ntc/ntc_service.cpp
@@ -93,7 +93,8 @@ bool NTC_Service::get_temperature(double &temperature){
     static unsigned int ntc_log_cnt = 0;
     if ( (ntc_log_cnt++ % 5u) == 0u ){
         unsigned int mv = (unsigned int)((adc_value * (unsigned int)NTC_VDDA_MV + (ADC_FULL_SCALE / 2u)) / ADC_FULL_SCALE);
-        LOG_INFO("NTC adc %u (%u mV), R %u ohm, temp %.1f C.", (unsigned int)adc_value, mv, (unsigned int)r_ntc, temperature);
+        LOG_INFO("NTC adc %u (%u mV), R %u ohm, temp %d C.",
+                 (unsigned int)adc_value, mv, (unsigned int)r_ntc, (int)temperature);
     }
     return true;
 }
